requirements.txt
fastapi
uvicorn
google-generativeai
chromadb
sentence-transformers
pydantic